# [Serving][Grammar] Integration of JSON schema generation